started the frontend

+++ backend/controllers/product.controller.js
@@ -54,12 +54,14 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
   
+  if(!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ success:false, message: "Product not found" });
+  }
   
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success:true, message: "Product deleted successfully" });
   } catch (error) {
-    res.status(400).json({ success:false, message: "Product not found" });
     return res.status(500).json({ success:false, message: "Server Error. Please try again." });
   }
 }
